services/hvac-analysis/core/document/table_extractor.py
"""
Table Extraction Module (Future Enhancement)
Specialized table detection and parsing for HVAC schedules

Status: Foundation/Stub for future implementation
Based on: Research Summary - Future Enhancements Section 2
"""

import cv2
import numpy as np
from typing import List, Dict, Optional, Tuple, Any
from dataclasses import dataclass
from enum import Enum
import logging

logger = logging.getLogger(__name__)


class TableType(Enum):
    """Types of tables in HVAC blueprints"""
    EQUIPMENT_SCHEDULE = "equipment_schedule"
    DUCT_SCHEDULE = "duct_schedule"
    PIPE_SCHEDULE = "pipe_schedule"
    DAMPER_SCHEDULE = "damper_schedule"
    DIFFUSER_SCHEDULE = "diffuser_schedule"
    LOAD_CALCULATION = "load_calculation"
    UNKNOWN = "unknown"


@dataclass
class Table:
    """Represents an extracted table"""
    bbox: Tuple[int, int, int, int]
    headers: List[str]
    rows: List[List[str]]
    table_type: TableType
    confidence: float
    metadata: Dict[str, Any]


class TableDetector:
    """
    Detect table regions in blueprint images
    
    TODO (Phase 1 - Months 1-3):
    - Implement line-based detection using Hough transform
    - Add ML-based detection (YOLOv8 fine-tuned on tables)
    - Support rotated tables
    """
    
    def __init__(self):
        logger.info("TableDetector initialized (stub)")
        
    def detect(self, image: np.ndarray) -> List[Dict[str, Any]]:
        """
        Detect table regions in image
        
        Returns:
            List of table regions with bounding boxes
            
        TODO:
        - Implement line detection
        - Cluster lines into table structures
        - Validate table candidates
        """
        logger.warning("TableDetector.detect() not yet implemented")
        return []


class TableStructureRecognizer:
    """
    Recognize table structure (rows, columns, cells)
    
    TODO (Phase 1):
    - Implement grid detection
    - Handle merged cells
    - Identify headers vs data rows
    """
    
    def __init__(self):
        logger.info("TableStructureRecognizer initialized (stub)")
        
    def analyze(self, table_region: np.ndarray) -> Dict[str, Any]:
        """
        Analyze table structure
        
        Returns:
            Structure information (rows, columns, cells)
            
        TODO:
        - Detect horizontal and vertical lines
        - Identify cell boundaries
        - Classify header rows
        """
        logger.warning("TableStructureRecognizer.analyze() not yet implemented")
        return {
            'rows': 0,
            'columns': 0,
            'cells': [],
            'headers': []
        }


class CellContentExtractor:
    """
    Extract content from table cells
    
    TODO (Phase 1):
    - ROI-based OCR per cell
    - Handle multi-line cells
    - Clean and normalize text
    """
    
    def __init__(self):
        logger.info("CellContentExtractor initialized (stub)")
        
    def extract(self, table_region: np.ndarray, 
               structure: Dict[str, Any]) -> List[List[str]]:
        """
        Extract text content from cells
        
        TODO:
        - Apply OCR to each cell region
        - Handle empty cells
        - Normalize formatting
        """
        logger.warning("CellContentExtractor.extract() not yet implemented")
        return [[]]


class TableExtractor:
    """
    Main table extraction pipeline
    
    Coordinates:
    - Table detection
    - Structure recognition  
    - Content extraction
    """
    
    def __init__(self):
        self.detector = TableDetector()
        self.structure_analyzer = TableStructureRecognizer()
        self.cell_extractor = CellContentExtractor()
        logger.info("TableExtractor initialized")
        
    def extract_tables(self, image: np.ndarray) -> List[Table]:
        """
        Extract all tables from image
        
        Pipeline:
        1. Detect table regions
        2. Analyze structure
        3. Extract cell contents
        4. Build table objects
        
        Args:
            image: Input blueprint image
            
        Returns:
            List of extracted Table objects
        """
        logger.info("Extracting tables from image")
        
        # Detect tables
        table_regions = self.detector.detect(image)
        logger.info(f"Detected {len(table_regions)} table regions")
        
        tables = []
        for region in table_regions:
            # Extract structure
            structure = self.structure_analyzer.analyze(region)
            
            # Extract contents
            cells = self.cell_extractor.extract(region, structure)
            
            # Build table
            table = self._build_table(region, structure, cells)
            tables.append(table)
        
        return tables
    
    def _build_table(self, region: Dict, structure: Dict, 
                    cells: List[List[str]]) -> Table:
        """Build Table object from extracted data"""
        return Table(
            bbox=(0, 0, 0, 0),  # TODO: Use actual bbox
            headers=[],  # TODO: Extract headers
            rows=cells,
            table_type=TableType.UNKNOWN,
            confidence=0.0,
            metadata={}
        )


class ScheduleRecognizer:
    """
    HVAC schedule recognition and parsing
    
    TODO (Phase 1 - Month 2):
    - Schedule type classification
    - Equipment parameter extraction
    - Cross-reference linking
    - Compliance validation
    """
    
    def __init__(self):
        logger.info("ScheduleRecognizer initialized (stub)")
        
    def recognize_schedule(self, table: Table) -> Dict[str, Any]:
        """
        Parse HVAC schedule from table
        
        TODO:
        - Identify schedule type
        - Extract equipment entries
        - Parse specifications
        - Link references
        """
        logger.warning("ScheduleRecognizer.recognize_schedule() not yet implemented")
        
        return {
            'schedule_type': TableType.UNKNOWN.value,
            'equipment': [],
            'specifications': {},
            'references': []
        }


class FormExtractor:
    """
    Extract structured data from forms
    
    TODO (Phase 1 - Month 3):
    - Template matching
    - Field detection
    - Value extraction
    - Validation
    """
    
    def __init__(self):
        self.templates = {}  # TODO: Load form templates
        logger.info("FormExtractor initialized (stub)")
        
    def extract_form(self, image: np.ndarray, form_type: str) -> Dict[str, Any]:
        """
        Extract structured data from form
        
        TODO:
        - Load appropriate template
        - Align form to template
        - Extract field values
        - Validate extracted data
        """
        logger.warning("FormExtractor.extract_form() not yet implemented")
        
        return {
            'form_type': form_type,
            'fields': {},
            'confidence': 0.0
        }


# Factory functions

def create_table_extractor() -> TableExtractor:
    """Create table extractor instance"""
    return TableExtractor()


def create_schedule_recognizer() -> ScheduleRecognizer:
    """Create schedule recognizer instance"""
    return ScheduleRecognizer()


def create_form_extractor() -> FormExtractor:
    """Create form extractor instance"""
    return FormExtractor()


# Example usage (for documentation)
if __name__ == "__main__":
    """
    Example usage of table extraction (when implemented):
    
    from core.document.table_extractor import create_table_extractor
    
    extractor = create_table_extractor()
    image = cv2.imread('hvac_blueprint.png')
    
    tables = extractor.extract_tables(image)
    
    for table in tables:
        print(f"Table Type: {table.table_type}")
        print(f"Headers: {table.headers}")
        print(f"Rows: {len(table.rows)}")
    """
    print("Table extraction module - stub implementation")
    print("See FUTURE_ENHANCEMENTS_ROADMAP.md for implementation plan")
