services/hvac-analysis/requirements.txt
# =============================================================================
# Core FastAPI dependencies
# =============================================================================
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart>=0.0.9
pydantic>=2.0.0

# =============================================================================
# AI/ML and Computer Vision
# =============================================================================
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv11 for object detection (bounding boxes)
transformers>=4.35.0  # Updated for Qwen2-VL support
diffusers>=0.21.0
accelerate>=0.24.0  # Updated for VLM training
bitsandbytes>=0.41.0  # For quantization (int8/int4)
peft>=0.7.0  # Parameter-Efficient Fine-Tuning (LoRA)
sentencepiece>=0.1.99  # Tokenization for VLMs
protobuf>=3.20.0  # For model serialization

# SAHI - Slicing Aided Hyper Inference for HVAC component detection
sahi>=0.11.0

# NetworkX for connectivity analysis (YOLOplan integration)
networkx>=3.1

# =============================================================================
# Image Processing and Computer Vision
# =============================================================================
opencv-python>=4.8.0  # Basic image processing only (no mask operations)
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.21.0
matplotlib>=3.7.0
# Note: pycocotools and SAM removed - using YOLOv11 object detection only

# =============================================================================
# Document Processing
# =============================================================================
pypdf2>=3.0.0
pymupdf>=1.23.0
pdf2image>=1.16.0
ezdxf>=1.1.0  # DWG/DXF processing
svgwrite>=1.4.0  # For synthetic data generation

# =============================================================================
# OCR (Optical Character Recognition)
# =============================================================================
pytesseract>=0.3.10
easyocr>=1.7.0
paddlepaddle>=2.5.0  # PaddlePaddle for PaddleOCR
paddleocr>=2.7.0     # PaddleOCR for advanced text extraction

# =============================================================================
# Data Processing
# =============================================================================
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# =============================================================================
# Graph Analysis and Spatial Indexing (for HVAC System Analysis)
# =============================================================================
networkx>=3.1  # System graph construction and analysis
rtree>=1.0.0   # R-tree spatial indexing for proximity queries

# =============================================================================
# Distributed Inference and Deployment
# =============================================================================
ray[serve]>=2.9.0  # Ray Serve for distributed inference graph

# =============================================================================
# Performance Optimization
# =============================================================================
python-dotenv>=1.0.0
tqdm>=4.65.0
redis>=5.0.0  # Caching (optional, for production)
celery>=5.3.0  # Background tasks (optional, for production)
deepspeed>=0.12.0  # Distributed training for VLM (optional)
wandb>=0.16.0  # Experiment tracking (optional)

# =============================================================================
# Utilities
# =============================================================================
requests>=2.31.0
aiohttp>=3.9.0
python-json-logger>=2.0.7

